+++ Tests/Loader/YamlFileLoaderTest.php
@@ -257,7 +257,7 @@ class YamlFileLoaderTest extends TestCase
         $this->assertTrue($def->hasTag('another.tag'));
         $this->assertSame([['foo' => 'bar']], $def->getTag('my.tag'));
         $this->assertSame([[]], $def->getTag('another.tag'));
-        $this->assertTrue($def->isAbstract());
+        $this->assertFalse($def->isAbstract());
     }
 
     public function testLoadShortSyntax()